### 2020-02-03 #2065 @whedon
[PRE REVIEW]: Expectations.jl: Quick and Accurate Expectation Operators in Julia
**Submitting author:** @arnavs (<a href="http://orcid.org/0000-0003-0074-7908">Arnav Sood</a>)
**Repository:** <a href="https://github.com/QuantEcon/Expectations.jl" target ="_blank">https://github.com/QuantEcon/Expectations.jl</a>
**Version:** v1.2.0
