### 2020-02-04 42bbfbd
add missing DOIs
