+++ paper.bib
@@ -4,6 +4,7 @@ author = {McCall, J. J.},
 year = {1970},
 journal = {The Quarterly Journal of Economics},
 volume = {84},
+doi = {10.2307/1879403},
 number = {1},
 pages = {113-126},
 abstract = {I. Introduction, 113. — II. A simple model of job search, 115. — III. A more general model of job search, 123. — IV. An adaptive search model, 125.},
@@ -16,7 +17,8 @@ url = {https://EconPapers.repec.org/RePEc:oup:qjecon:v:84:y:1970:i:1:p:113-126.}
     year={2014},
     eprint={1410.5286},
     archivePrefix={arXiv},
-    primaryClass={math.NA}
+    primaryClass={math.NA},
+    doi = {10.1093/imanum/drv002}
 }
 
 @misc{distributions,
