### 2020-02-04 #2065 @kthyng
@VivianePons Can you take a look at this submission? I'd ask you to be the editor if we take it on.

@arnavs Thank you for your submission! The editorial board at JOSS is currently working on improved guidelines for in-scope submissions. We haven't finished, but we plan to start using some metrics to help guide if a submission passes some minimum bars to be considered a "large enough" contribution to be considered for JOSS, of course a question at any journal. Your submission is a bit on the low end for lines of code. This is not disqualifying, but it does mean that it would help if you were to give some extra motivation and understanding for the editors here who are not necessarily experts in your domain as to why your submission is a large enough contribution for its own publication. Thanks.
