### 2020-02-04 #2065 @kthyng
@arnavs Yes this would be excellent to add to the paper, but also please provide the information here when you have it in case there is some discussion to be had.
