### 2020-02-04 #2065 @arnavs
@kthyng Sure. So the crux is that many statistical problems require taking an expectation of some random variable. This comes up all the time in various domains (economics, chemistry, physics, etc.)

Julia has really great support for distribution objects. But given one of those distribution objects (say, `d = Normal()`), there's really no easy way (besides drawing a large sample and averaging) to answer questions like "what is the expected value of `sin(x)` where `x` is drawn from `d`?)

What we do in this package is create operators `E = expectation(d)`, where you can go `E(x -> sin(x))`, etc. The operators themselves are computed with Gaussian quadrature (specialized for various distributions, like `Exponential`, `Erlang`, `Beta`, etc.) We referred to @pkofod's `DistQuads.jl` library for the underlying calls to the quadrature library.

These operators are more efficient, reusable, notationally clear, and highly accurate. Instead of taking a sample of `10^6` to compute an expectation, you can just multiply two vectors of size 32.

Hope this helps.
