### 2020-02-05 #2065 @VivianePons
@kthyng From what I understand, the code, even if not big, is quite meaningful and implements some needed mathematics. I believe we can consider it for submission and move forward. Of course, the final decision will depend on the reviewers comments. Do you agree?
