### 2020-02-06 #2065 @arnavs
@VivianePons We had a recent request from a user for the `Erlang` distribution, so I added that in. The package is used in the [QuantEcon lecture notes](julia.quantecon.org), and was also used as a dependency for a package `AdaptiveGaussianProcesses.jl` (although I think it may not be used in the latest version.)

Will add this to the paper. thanks for the suggestion.
