### 2020-02-07 #2065 @VivianePons
Hi @pkofod you were suggested as a potential reviewer for this submission into JOSS (The Journal of Open Source Software), would you be interested in carrying it out?

JOSS is a free, open-source, community driven and developer-friendly online journal
(no publisher is seeking to raise revenue from the volunteer labor of researchers!).

The review process at JOSS is unique: it is open and author-reviewer-editor conversations
are encouraged.

JOSS reviews involve downloading and installing the software, and inspecting the repository
and submitted paper for key elements. See https://joss.readthedocs.io/en/latest/review_criteria.html

Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to
satisfy the review).

Thank you for your consideration!
