+++ Project.toml
@@ -1,6 +1,6 @@
 name = "Expectations"
 uuid = "2fe49d83-0758-5602-8f54-1f90ad0d522b"
-version = "1.2.0"
+version = "1.3.0"
 
 [deps]
 Compat = "34da2185-b29b-5c13-b0c7-acf172513d20"
