### 2020-02-11 #2090 @odow
Only one point with the paper: you should reference `FastGaussQuadrature.jl`.

The real go/no-go question is the extent to which the library is substantive enough. I note that this was [brought up in pre-review](https://github.com/openjournals/joss-reviews/issues/2065#issuecomment-582000457):

> The editorial board at JOSS is currently working on improved guidelines for in-scope submissions. We haven't finished, but we plan to start using some metrics to help guide if a submission passes some minimum bars to be considered a "large enough" contribution to be considered for JOSS, of course a question at any journal. Your submission is a bit on the low end for lines of code. This is not disqualifying, but it does mean that it would help if you were to give some extra motivation and understanding for the editors here who are not necessarily experts in your domain as to why your submission is a large enough contribution for its own publication.

I would tend to agree.

`Expectations.jl` is on the low-side with respect to lines of code, and the real difficulty is handled by the `FastGaussQuadrature.jl` package. In addition, `FastGaussQuadrature.jl` doesn't meet the JOSS standards: CI appears to be down, limited documentation, etc.

I'll defer to an editor on the decision though; it does seem shame to place barriers in front of people trying to get credit for their open-source work.

If the `FastGaussQuadrature.jl` package was tidied up, it would be a suitable JOSS paper, and then you could say "oh, and there is a package called Expectations.jl that uses FastGaussQuadrature.jl to do cool things."
