### 2020-02-11 #2090 @arnavs
@odow Can you elaborate on CI? Last I checked (this morning), everything is passing on Travis and AppVeyor.

Documentation is a bit thin, but the basic functionality seems pretty straightforward. Would be happy to elaborate if you have any suggestions.

We cite FastGaussQuadrature in the paper. Will search around for a formal thing to refer to.
