### 2020-02-12 #2090 @arnavs
@odow I asked the FGQ guys, and the paper I cite (Townsend, Olver, Trogdon) is what they recommend citing. I think that + the inline reference to FGQ should do it.
