### 2020-02-16 2df81f6
add Uniform and update docs, tests
