+++ test/iterable.jl
@@ -9,6 +9,7 @@ distset = [
     LogNormal(4.5),
     DiscreteUniform(1, 10),
     Chisq(3),
+    Uniform(1, 200),
     Binomial(101, 0.45),
     Categorical([0.2, 0.35, 0.15, 0.3])
 ]
