+++ README.md
@@ -11,7 +11,7 @@ Installation (v1.0 and up):
 pkg> add Expectations
 ```
 
-This is a package designed to simplify the process of taking expectations of functions of random variables. 
+This is a package designed to simplify the process of taking expectations of functions of random variables.
 
 ### Expectation Operator
 
@@ -36,7 +36,7 @@ f(x) = x^2
 E * f.(x) == dot(f.(x), weights(E))
 ```
 
-### Random Variables 
+### Random Variables
 
 The underlying distributions are objects from `Distributions.jl` (currently `<:UnivariateDistribution`).
 
@@ -44,16 +44,17 @@ The underlying distributions are objects from `Distributions.jl` (currently `<:U
 
 We support different types of Gaussian quadrature (Gauss-Hermite, Gauss-Legendre, Gauss-Laguerre, etc.) based on the distribution, as well as some methods with user-defined nodes (e.g., trapezoidal integration).
 
-We have rules for the following distributions: 
+We have rules for the following distributions:
 
-* Normal 
+* Normal
 * ChiSq
 * LogNormal
 * Exponential
 * Beta
 * Gamma/Erlang
+* Uniform
 * Continuous Univariate (compact; generic fallback)
 * Continuous Univariate (no restriction; approximates with quantile grid)
 * Discrete
 
-See docs for more info. 
+See docs for more info.
