### 2020-02-16 3adbc7e
remove travis documenter [ci skip]
