### 2020-02-21 #2090 @oxinabox
I have not yet properly begun the review, and am yet to more than skim the paper and code.

I notice that a few lines are almost character for character identical to DistQuads.jl
The DistQuads package is mentioned in Acknowledgements.
However, given relative size and scope of the overlap, it seems as though it may be more appropriate to have the author of DistQuads.jl as a coauthor to the paper.
e.g.:
https://github.com/QuantEcon/Expectations.jl/blob/0a24f4b53ed2bc4fc6119a0a2b69c550a571dab3/src/iterable.jl#L175
and
https://github.com/pkofod/DistQuads.jl/blob/553a1a7dcf6fd679aae7f7bd7c19f0b511040edd/src/DistQuads.jl#L46
and several other lines from that file.
