### 2020-02-21 #2090 @arnavs
@oxinabox I ran the acknowledgment text by Patrick (and this is kinda why I wanted him to be a reviewer.) But that's a good point. I'll ask him if he's interested in coauthorship. Will let you know what he says.
