### 2020-02-21 0b5493c
remove acknowledgmeent and reorder affililations
