+++ src/Expectations.jl
@@ -2,17 +2,19 @@
 
 module Expectations
 
-# Load external dependencies. 
+# Load external dependencies.
 using Compat, FastGaussQuadrature
 using Compat.LinearAlgebra
 using SpecialFunctions
 using Distributions
 
-# Load internal files. 
+# Load internal files.
 include("types.jl"),
 include("iterable.jl")
+include("mixturemodels.jl")
 
-# Export 
+# Export
 export expectation, Expectation, IterableExpectation, Gaussian, Trapezoidal, FiniteDiscrete, QuadratureAlgorithm, ExplicitQuadratureAlgorithm, nodes, weights, QuantileRange
+export MixtureExpectation
 
 end # module
