+++ src/mixturemodels.jl
@@ -13,4 +13,4 @@ weights(e::MixtureExpectation) = e.mixtureweights
 expectations(e::MixtureExpectation) = e.expectations
 expectation(f::Function, m::UnivariateMixture; kwargs...) = dot(probs(m), [expectation(f, dist; kwargs...) for dist in components(m)])
 *(r::Real, e::MixtureExpectation) = MixtureExpectation(r*expectations(e), weights(e))
-*(e::MixtureExpectation, h::AbstractArray) = dot(weights(e), [E*h for E in expectations(e)])
+# *(e::MixtureExpectation, h::AbstractArray) = dot(weights(e), [E*h for E in expectations(e)])
