### 2020-02-29 #2090 @oxinabox
Below follows my, admittedly fairly limited, evaluation of performance.

```julia
julia> using Distributions, Expectations, BenchmarkTools

julia> const y = Normal(20, 0.01)
Normal{Float64}(μ=20.0, σ=0.01)

julia> foo(x) = x^2 - 0.5x
foo (generic function with 1 method)

julia> @btime mean(foo, rand($y, 1000))
  9.807 μs (1 allocation: 7.94 KiB)
390.0042603170717

julia> @btime mean(foo, rand($y, 1_000_000))
  10.538 ms (2 allocations: 7.63 MiB)
389.9998149984719

julia> @btime expectation(foo, $y)
  20.196 μs (254 allocations: 31.38 KiB)
390.0000999999998
```

Seems it is much faster than working with a large number of samples, but similarly accurate.

Seems good
