### 2020-02-29 #2090 @oxinabox
>State of the field: Do the authors describe how this software compares to other commonly-used packages?

It would be nice to see reference to other software.
I am sure they must have tools for something similar in R or Python.

---

It would also be nice to see some examples in the documentation of how it's generic fallback can be used with other distribution types.
E.g. with [TransformedDistributions from Bijectors.jl](https://github.com/TuringLang/Bijectors.jl).

More broadly, and related to this, it would be good to contextualize how this fits into the world of PPLs.
