### 2020-02-29 #2090 @arnavs
@oxinabox Thanks for finishing the review. And good idea. Will add some references to other langs and types.

There's also the classic `Measurements.jl` add-on that works super well.

```
julia> d = Normal(1. ± 0.5, 0.25 ± 0.25);

julia> E = expectation(d);

julia> E(x -> x^2 * sin(x))
0.9 ± 1.1
```
