### 2020-03-07 #2090 @arnavs
@oxinabox Incorporated your feedback; the TransformedDistribution support is looking a bit more annoying than I imagined (partly bc of some missing methods, partly because I need to read the math.) So I filed an issue, noted what our generic fallback supports, and will work on extending that.
