### 2020-03-07 d19b3f5
change unicode to mu, sigma
