### 2020-03-07 #2090 @arnavs
(unicode support fix)
