### 2020-03-09 #2090 @VivianePons
Hi @arnavs thank you for incorporating all the feedbacks. I am still feeling some hesitations from the reviewers. I would like to have their opinions now that many things have changed, especially a new author!

Dear @odow and @oxinabox thank you very much for the work you have provided. I see that @arnavs has implemented some of your suggestions. How do you feel about the new version of the software / paper, especially the clarification on references and the addition of a new author? would you say that this piece is a "significant contribution" to the existent state of scientific open source software on this topic?
