### 2020-03-09 #2090 @oxinabox
Without the addition of the new author, I would have said this fails the requirements for "Contribution and authorship".

I think this probably marks the low-end of _" "significant contribution" to the existent state of scientific open source software on this topic?"_.
Calculated expectations is important for many things, and this makes it faster and easier.
