### 2020-03-09 #2090 @odow
One typo: `s/elswhere/elsewhere`.

Agree with @oxinabox. Low-end of significant contribution, but still useful.

This is really an editorial decision on where the bar should be set with respect to papers accepted.

If the decision is to have a generous acceptance policy in order to give people credit for their open-source work (Which we appreciate, @arnavs!), accept.
If JOSS wants to move toward a more stringent policy that selectively rewards more significant contributions, reject.
