### 2020-03-09 #2090 @arnavs
Good catch on typo @odow. Fixed.

And thanks @VivianePons. Just a note that it would be great to re-check the statistics if possible. I don't know if we are over the threshold, but we have been adding features here and there (new distributions, mixed distributions, etc.)
