### 2020-03-13 #2090 @VivianePons
Dear @arnavs thank you very much for your patience.

We have had a thorough discussion with the editorial team. I am sorry but we believe that the reviewer comments confirm our initial concern that the submission is a bit too small to be considered for publication (the quantity of code is an indicator here but not a barrier in itself) . We are right now improving our processes and guidelines to help authors better understand the expectations of the journal and improve the handling of submission like yours which are on the edge: https://github.com/openjournals/joss/issues/709

Nevertheless, we value the high quality work that you have done and which was noted by the reviewers and the editorial team. If, in the future, you add  significant features to your software, we would suggest that you resubmit to be again considered for publication.

I understand that this decision might come as disappointing but I hope this won't discourage any further work of yours. I have very much appreciated your responsiveness dealing with the editors and reviewers comments.

I wish you the best and long life to Expectations.jl and hopefully a publication in JOSS when it has grown a bit.
